+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Changed hierarchy of Secondary Button of the IFPE Modal
+
 ## [4.60.0] - 2022-05-05Z
 
 ### Added