+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+- Added BIN setter in Tokenization services
 - Changed hierarchy of Secondary Button of the IFPE Modal
 
 ## [4.60.0] - 2022-05-05Z