+++ CHANGELOG.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [4.63.0] - 2022-05-19Z
 ### Fixed
 - Fixed Analytics Tracks for Remedies
+- Fixed sonic branding resource path
 
 ## [4.62.0] - 2022-05-12Z
 - Added Reauth support